--- a/migration/bucket_migrator.go
+++ b/migration/bucket_migrator.go
@@ -3,9 +3,12 @@ package migration
 import (
 	"bytes"
 	"context"
+	"fmt"
+	"runtime"
 
 	"github.com/NethermindEth/juno/db"
 	"github.com/NethermindEth/juno/utils"
+	"github.com/sourcegraph/conc/pool"
 )
 
 var _ Migration = (*BucketMigrator)(nil)
@@ -76,34 +79,62 @@ func (m *BucketMigrator) Migrate(_ context.Context, txn db.Transaction, network 
 	remainingInBatch := m.batchSize
 	iterator, err := txn.NewIterator()
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("new iterator: %v", err)
 	}
 
+	keyValue := make([][2][]byte, 0)
+	callWithNewTransaction := false
 	for iterator.Seek(m.startFrom); iterator.Valid(); iterator.Next() {
 		key := iterator.Key()
 		if !bytes.HasPrefix(key, m.target.Key()) {
 			break
 		}
-
-		if pass, err := m.keyFilter(key); err != nil {
-			return nil, utils.RunAndWrapOnError(iterator.Close, err)
-		} else if pass {
-			if remainingInBatch == 0 {
-				m.startFrom = key
-				return nil, utils.RunAndWrapOnError(iterator.Close, ErrCallWithNewTransaction)
-			}
-
-			remainingInBatch--
-			value, err := iterator.Value()
-			if err != nil {
-				return nil, utils.RunAndWrapOnError(iterator.Close, err)
-			}
-
-			if err = m.do(txn, key, value, network); err != nil {
-				return nil, utils.RunAndWrapOnError(iterator.Close, err)
-			}
+		if do, filterErr := m.keyFilter(key); filterErr != nil {
+			return nil, fmt.Errorf("filter: %v", filterErr)
+		} else if !do {
+			continue
 		}
+		if remainingInBatch == 0 {
+			m.startFrom = key
+			callWithNewTransaction = true
+			break
+		}
+		remainingInBatch--
+
+		var value []byte
+		value, err = iterator.Value()
+		if err != nil {
+			return nil, utils.RunAndWrapOnError(iterator.Close, fmt.Errorf("get value from iterator: %v", err))
+		}
+		keyValue = append(keyValue, [2][]byte{key, value})
+	}
+	if err = iterator.Close(); err != nil {
+		return nil, fmt.Errorf("close iterator: %v", err)
 	}
 
-	return nil, iterator.Close()
+	syncTxn := db.NewSyncTransaction(txn)
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+	p := pool.New().WithErrors().WithFirstError().WithMaxGoroutines(3 * runtime.GOMAXPROCS(0))
+	for _, kv := range keyValue {
+		if ctx.Err() != nil {
+			break
+		}
+		kv := kv
+		p.Go(func() error {
+			if doErr := m.do(syncTxn, kv[0], kv[1], network); doErr != nil {
+				cancel()
+				return fmt.Errorf("do: %v", doErr)
+			}
+			return nil
+		})
+	}
+	if err = p.Wait(); err != nil {
+		return nil, fmt.Errorf("worker: %v", err)
+	}
+
+	if callWithNewTransaction {
+		return nil, ErrCallWithNewTransaction
+	}
+	return nil, nil
 }