Memory usage roughly doubles because we build the list of keys and values from the iterator. 

If this is a problem, we could break up the batch into smaller batches and parallelize each individually, giving the garbage collector a chance to clean up some of the memory.

Before

```
% go test -bench=. -benchtime=20s      
BenchmarkBucketMigratorMemDB-12    	    302	 82_671_637 ns/op

% go test -bench=. -benchtime=20s -benchmem
BenchmarkBucketMigratorMemDB-12    	    303	 79092378 ns/op	44442478 B/op	 200_507 allocs/op
```

After

```
% go test -bench=. -benchtime=20s
BenchmarkBucketMigratorMemDB-12              160         147_339_162 ns/op

% go test -bench=. -benchtime=20s -benchmem
BenchmarkBucketMigratorMemDB-12              160         165104490 ns/op        82721453 B/op     400_645 allocs/op
```